requirements.txt
PyGithub==1.54.1
python-dotenv==0.15.0
slack-sdk==3.2.0